# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sound2Soul API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /creators/onboarding:
    post:
      operationId: createCreatorProfile
      tags: [creators]
      summary: Create creator profile (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreatorProfileBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorProfile"
  /creators/me:
    get:
      operationId: getMyCreatorProfile
      tags: [creators]
      summary: Get own creator profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorProfile"
    patch:
      operationId: updateCreatorProfile
      tags: [creators]
      summary: Update own creator profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCreatorProfileBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorProfile"
  /creators/{slug}:
    get:
      operationId: getCreatorBySlug
      tags: [creators]
      summary: Get public creator profile by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorPublicProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /creators/{slug}/tracks:
    get:
      operationId: getCreatorTracks
      tags: [creators]
      summary: Get public tracks by creator slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"

  # Tracks
...
  /likes/{trackId}:
    post:
      operationId: likeTrack
      tags: [interactions]
      summary: Like a track
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: unlikeTrack
      tags: [interactions]
      summary: Unlike a track
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /saves/{trackId}:
    post:
      operationId: saveTrack
      tags: [interactions]
      summary: Save a track
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: unsaveTrack
      tags: [interactions]
      summary: Unsave a track
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /follows/{creatorId}:
    post:
      operationId: followCreator
      tags: [interactions]
      summary: Follow a creator
      parameters:
        - name: creatorId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: unfollowCreator
      tags: [interactions]
      summary: Unfollow a creator
      parameters:
        - name: creatorId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /interactions/{trackId}:
    get:
      operationId: getTrackInteractions
      tags: [interactions]
      summary: Get current user's interaction state for a track
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackInteractionState"

  # Fan emails
...
  /reports:
    post:
      operationId: reportTrack
      tags: [reports]
      summary: Report a track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Waitlist
...
  /admin/reports:
    get:
      operationId: getAdminReports
      tags: [admin]
      summary: Get all reports
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, reviewed, resolved]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    CreatorProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        artistName:
          type: string
        slug:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        bannerUrl:
          type: string
        genres:
          type: array
          items:
            type: string
        moodIdentityTags:
          type: array
          items:
            type: string
        aiToolsUsed:
          type: array
          items:
            type: string
        socialLinks:
          type: object
          additionalProperties:
            type: string
        creatorStatement:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, artistName, slug]
    CreatorPublicProfile:
      type: object
      properties:
        id:
          type: string
        artistName:
          type: string
        slug:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        bannerUrl:
          type: string
        genres:
          type: array
          items:
            type: string
        moodIdentityTags:
          type: array
          items:
            type: string
        aiToolsUsed:
          type: array
          items:
            type: string
        socialLinks:
          type: object
          additionalProperties:
            type: string
        creatorStatement:
          type: string
        followerCount:
          type: integer
        trackCount:
          type: integer
        createdAt:
          type: string
      required: [id, artistName, slug]
    CreateCreatorProfileBody:
      type: object
      properties:
        artistName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        bannerUrl:
          type: string
        genres:
          type: array
          items:
            type: string
        moodIdentityTags:
          type: array
          items:
            type: string
        aiToolsUsed:
          type: array
          items:
            type: string
        socialLinks:
          type: object
          additionalProperties:
            type: string
        creatorStatement:
          type: string
      required: [artistName]
    UpdateCreatorProfileBody:
      type: object
      properties:
        artistName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        bannerUrl:
          type: string
        genres:
          type: array
          items:
            type: string
        moodIdentityTags:
          type: array
          items:
            type: string
        aiToolsUsed:
          type: array
          items:
            type: string
        socialLinks:
          type: object
          additionalProperties:
            type: string
        creatorStatement:
          type: string
    Track:
      type: object
      properties:
        id:
          type: string
        creatorId:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        audioUrl:
          type: string
        coverImageUrl:
          type: string
        genre:
          type: string
        moodTags:
          type: array
          items:
            type: string
        soulStory:
          type: string
        aiInvolvementType:
          type: string
        humanContributionChecklist:
          type: object
        rightsConfirmation:
          type: object
        visibility:
          type: string
          enum: [draft, public, unlisted]
        moderationStatus:
          type: string
          enum: [pending, approved, rejected]
        isFeatured:
          type: boolean
        playCount:
          type: integer
        likeCount:
          type: integer
        saveCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        creator:
          $ref: "#/components/schemas/CreatorPublicProfile"
      required: [id, creatorId, title, slug, visibility, moderationStatus]
...
    TrackInteractionState:
      type: object
      properties:
        isLiked:
          type: boolean
        isSaved:
          type: boolean
      required: [isLiked, isSaved]
...
    ReportBody:
      type: object
      properties:
        trackId:
          type: string
        reason:
          type: string
          enum: [copyright_concern, unauthorized_voice, offensive_content, misleading_disclosure, other]
        details:
          type: string
        reporterEmail:
          type: string
      required: [trackId, reason]
...
    Report:
      type: object
      properties:
        id:
          type: string
        trackId:
          type: string
        trackTitle:
          type: string
        trackSlug:
          type: string
        reporterEmail:
          type: string
        reason:
          type: string
        details:
          type: string
        status:
          type: string
          enum: [open, reviewed, resolved]
        createdAt:
          type: string
      required: [id, trackId, reason, status, createdAt]
...